requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# HTTP & Crawling
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.27.2

# LLM & Agent
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
langgraph==0.2.45
openai==1.52.2

# Vector DB
chromadb==0.5.18

# Evaluation & Observability
mlflow==2.17.2
langsmith==0.1.136

# Pipeline
apache-airflow==2.9.3

# Utilities
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.27.2
